backend/requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9
pydantic-settings>=2.2.1

# Scientific Computing
numpy
scipy

# Database & Vector Store
sqlalchemy[asyncio]>=2.0.29
aiosqlite>=0.20.0
chromadb>=0.4.24

# AI & Media Processing
funasr>=1.0.26
modelscope>=1.13.0
librosa>=0.10.1
google-genai>=0.6.0
llama-index>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
celery>=5.3.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.27.0
